+++ app/src/main/java/com/example/herbario_nacional/models/PlantSpecimen.kt
@@ -13,8 +13,8 @@ data class PlantSpecimen(
     val photo: String?,
     val photo_url: String?,
     val date_received: String,
-    // val family: PlantFamily,
-    // val genus: Genus,
+//     val family: PlantFamily,
+//     val genus: Genus,
     val species: PlantSpecies,
     val status: Status,
     val number_of_samples: Int,
