+++ app/src/main/java/com/example/herbario_nacional/ui/Fragments/SearchFragment.kt
@@ -25,10 +25,6 @@ class SearchFragment : Fragment() {
     private lateinit var tabs: TabLayout
     private lateinit var searchPagerAdapter: SearchAdapter
 
-    private var position: Int = 0
-    private var query: String = String()
-    private val searchViewModel: SearchViewModel by viewModel()
-
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
 
         val view: View = inflater.inflate(R.layout.fragment_search, container, false)
@@ -42,74 +38,4 @@ class SearchFragment : Fragment() {
 
         return view
     }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        // get current tab
-        tabs.addOnTabSelectedListener(object : OnTabSelectedListener {
-            override fun onTabReselected(tab: TabLayout.Tab?) {
-                val pass: Unit = Unit
-            }
-
-            override fun onTabUnselected(tab: TabLayout.Tab?) {
-                val pass: Unit = Unit
-            }
-
-            override fun onTabSelected(tab: TabLayout.Tab) {
-                position = tab.position
-            }
-        })
-
-        mi_search_view.setOnQueryTextListener(object: SearchView.OnQueryTextListener {
-
-            override fun onQueryTextSubmit(s:String):Boolean {
-
-                if(position == 0) {
-                    searchViewModel.searchPlantByName(s)
-                } else {
-                    searchViewModel.searchFungusByName(s)
-                }
-
-                Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
-                return false
-            }
-
-            override fun onQueryTextChange(s:String):Boolean {
-                query = s
-                return true
-            }
-        })
-
-          // button listeners
-        btn_common_name.setOnClickListener {
-            if(position == 0) {
-                searchViewModel.searchPlantByName(query)
-            } else {
-                searchViewModel.searchFungusByName(query)
-            }
-            Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
-        }
-
-        btn_location.setOnClickListener {
-            if(position == 0) {
-                searchViewModel.searchPlantByLocation(query)
-            } else {
-                searchViewModel.searchFungusByLocation(query)
-            }
-            Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
-        }
-
-        btn_recollection.setOnClickListener {
-            if(position == 0) {
-                searchViewModel.searchPlantByRecollectionArea(query)
-            } else {
-                searchViewModel.searchFungusByRecollectionArea(query)
-            }
-            Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
-        }
-
-    }
-
-
 }
