+++ app/src/main/java/com/example/herbario_nacional/ui/Fragments/SearchPlantFragment.kt
@@ -4,18 +4,16 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.SearchView
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
-import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.example.herbario_nacional.R
 import com.example.herbario_nacional.adapters.PlantAdapter
 import com.example.herbario_nacional.imageloader.ImageLoader
 import com.example.herbario_nacional.ui.viewModels.SearchViewModel
 import com.example.herbario_nacional.util.GridItemDecoration
-import com.facebook.shimmer.ShimmerFrameLayout
-import kotlinx.android.synthetic.main.fragment_plants.*
 import kotlinx.android.synthetic.main.fragment_search_plants.*
 import org.koin.android.ext.android.inject
 import org.koin.androidx.viewmodel.ext.android.viewModel
@@ -25,6 +23,8 @@ class SearchPlantFragment : Fragment() {
     private val imageLoader: ImageLoader by inject()
     private val plantAdapter: PlantAdapter by lazy { PlantAdapter(imageLoader) }
     private val searchViewModel: SearchViewModel by viewModel()
+
+    private var query: String = String()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_search_plants, container, false)
@@ -39,7 +39,6 @@ class SearchPlantFragment : Fragment() {
             val dataState = it ?: return@Observer
             if (dataState.result != null && !dataState.result.consumed){
                 dataState.result.consume()?.let { result ->
-                    println(result)
                     plantAdapter.submitList(result)
                 }
             }
@@ -49,6 +48,35 @@ class SearchPlantFragment : Fragment() {
                 }
             }
         })
+
+        plant_search_view.setOnQueryTextListener(object: SearchView.OnQueryTextListener {
+
+            override fun onQueryTextSubmit(s:String):Boolean {
+                searchViewModel.searchPlantByName(s)
+                Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
+                return false
+            }
+
+            override fun onQueryTextChange(s:String):Boolean {
+                query = s
+                return true
+            }
+        })
+
+        btn_common_name.setOnClickListener {
+            searchViewModel.searchPlantByName(query)
+            Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
+        }
+
+        btn_location.setOnClickListener {
+            searchViewModel.searchPlantByLocation(query)
+            Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
+        }
+
+        btn_recollection.setOnClickListener {
+            searchViewModel.searchPlantByRecollectionArea(query)
+            Toast.makeText(context, getString(R.string.searching), Toast.LENGTH_SHORT).show()
+        }
     }
 
     private fun setupRecycler(){
